requirements.txt
transformers==2.10.0
pytorch-crf==0.7.2
torch==1.5.0
tqdm>=4.46.0
numpy==1.18.1
grpcio==1.38.1             
grpcio-tools==1.38.1
tensorboard==2.3.0
python-Levenshtein==0.12.2             
